langgraph/requirements.txt
langgraph
langchain-core
langchain-openai
sochdb>=0.4.0
python-dotenv
numpy
